Fixed pagination bug; passed validation
This commit fixes the following problems:

1) Previously, each time the script parsed a new page of API results
from CONTENTdm, it would re-initialize the collection_blob
data structure, obliterating the previous pages of manifests.
This would have the effect of creating a collection manifest file
that only contained the final page of data.

The collection_blob data structure has been made global, and
initialized only once per execution.

2) The collection manifest failed validation against the
presentation API because of a non-URI `@id`.  Changing the `@id`
to the semantically bogus (but syntactically valid)
`http://example.com/GUID_GOES_HERE` allows the manifest to pass
validation.

3) The collection manifest failed validation because of a lack
of a label.  We are now generating a label.